+++ resources/js/admin/view/main/reviews/ReviewsReported.vue
@@ -180,6 +180,7 @@
             // console.log(this.itemsData[this.selectedIndex])
           })
           .catch((error) => {
+            this.toastError(error.response.data.message)
             console.log(error);
           })