+++ resources/js/admin/view/main/reviews/ReviewsSuspended.vue
@@ -178,6 +178,8 @@
             // console.log(this.itemsData[this.selectedIndex])
           })
           .catch((error) => {
+            this.toastError(error.response.data.message)
+
             console.log(error);
           })