+++ resources/js/website/components/CompanyRegister.vue
@@ -2,7 +2,10 @@
   <div class="h-100 position-relative">
     <div id="login-form" class="text-center position-absolute bg-white h-100 px-3" style="right:0">
       <div class=" py-3 border-bottom">
-        <h3 class="mb-0">APP Logo</h3>
+        <a href="/">
+          <b-img style="max-height:30px" src="/images/logo-large.png" alt="application logo"></b-img>
+        </a>
+
       </div>
       <div class="mt-3">
         <h4>Business Account Register</h4>