+++ resources/views/layouts/admin.blade.php
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="apple-touch-icon" sizes="180x180" href="{{ asset('images/apple-touch-icon.png') }}">
-  <link rel="icon" type="image/x-icon" href="{{ asset('images/favicon.ico') }}">
+  <link rel="icon" type="image/x-icon" href="{{ asset('images/favicon.png') }}">
   <link rel="manifest" href="{{ asset('images/site.webmanifest') }}">
   <!-- Main Stylesheet File -->
   {{-- <link href="{{ mix('css/app.css') }}" rel="stylesheet"> --}}