+++ resources/views/layouts/website.blade.php
@@ -10,7 +10,7 @@
   <meta name="webcrawlers" content="all">
   <meta name="theme-color" content="#041d4c" />
   <link rel="apple-touch-icon" sizes="180x180" href="{{ asset('images/apple-touch-icon.png') }}">
-  <link rel="icon" type="image/x-icon" href="{{ asset('images/favicon.ico') }}">
+  <link rel="icon" type="image/x-icon" href="{{ asset('images/favicon.png') }}">
   <link rel="manifest" href="{{ asset('images/site.webmanifest') }}">
   <!-- Main Stylesheet File -->
   {{-- <link href="{{ mix('css/app.css') }}" rel="stylesheet"> --}}