+++ resources/views/pages/user-register.blade.php
@@ -9,8 +9,9 @@
   <div class="d-flex justify-content-center align-items-center h-100 px-3">
     <div id="login-form" class="text-center card  p-3">
       <div class=" pb-3 border-bottom">
-        <h3 class="mb-0">APP Logo</h3>
-      </div>
+  <a href="/">
+    <img class="img-fluid" style="max-height:30px" src="/images/logo-large.png" alt="application logo"></img>
+  </a> </div>
       <div class=" py-2">
         <a href="{{ url('facebook') }}" class="btn btn-block text-white" style="background-color: #3B5998;">
           <i class="fa fa-facebook"> </i>